systemd-resolved: add ResetDNSServers method to Manager interface
When a lookup fails using CurrentDNServer, resolved will reattempt the query with
each server defined in the DNS property in turn until one is successful (or they
all fail). The first successful server is assigned to the CurrentDNSServer
property and used first for future queries. Short of restarting resolved, there
is no way for the user to force resolved to attempt to use the first server again.

ResetDNSServers is a method on the Manager and Link interfaces that when called will
force resolved to set CurrentDNSServer property to the first DNS server defined
in the DNS property.