systemd-resolved: add dns server reset timer and config
When a lookup fails using CurrentDNServer, resolved will reattempt the query with
each server defined in the DNS property in turn until one is successful (or they
all fail). The first successful server is assigned to the CurrentDNSServer
property and used first for subsequent queries.

This patch introduces a timer that is started after the first server fails in
a transaction. The timeout value is defined in seconds and configured with
DNSServerResetSec. After the timer is triggered, it will attempt to reset the
CurrentDNSServer property to the first entry in the DNS server list. The
default is 0 which means there is no reset timer set.